+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Made `change_user_role` public
 * Added `no_course_lessons_id_zero()`, so lessons not assigned to a course get lesson ID `0` stored, to enable easier selection of lessons with no course.
 * Added `pcm_no_course_lesson_slug` filter.
+* Added `user_has_booked_course` method.
 
 ## 0.2.2
 * More testing